requirements.txt
Flask==3.1.2
flask-cors==6.0.2
pymongo==4.16.0
PyJWT==2.12.1
python-dotenv==1.2.2
bcrypt==5.0.0
certifi==2025.10.5
dnspython==2.8.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.0
MarkupSafe==3.0.2
blinker==1.9.0
requests==2.32.5
urllib3==2.5.0
charset-normalizer==3.4.4
idna==3.11
certifi==2025.10.5
